Fixed value type iris
